server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.12.0
numpy>=2.0.0
networkx>=3.0
gymnasium>=1.0.0
stable-baselines3>=2.8.0
torch>=2.3.0
openai>=2.0.0
